move str_to_class to utils
