--- deepinv/optim/__init__.py
@@ -1,4 +1,0 @@
-import sys 
-from deepinv.optim.optim_iterators import *
-def str_to_class(classname):
-    return getattr(sys.modules[__name__], classname)
