+++ deepinv/optim/optim_iterators/optim_iterator.py
@@ -1,6 +1,5 @@
 import torch
 import torch.nn as nn
-from deepinv.optim.utils import gradient_descent
 
 
 class OptimIterator(nn.Module):
@@ -70,10 +69,11 @@ class gStep(nn.Module):
 
         if prox_g is None and grad_g is None:
             if g is not None and isinstance(g, nn.Module):
+                
                 def grad_g(self, x, *args):
                     torch.set_grad_enabled(True)
                     return torch.autograd.grad(g(x, *args), x, create_graph=True, only_inputs=True)[0]
-
+                from deepinv.optim.utils import gradient_descent
                 def prox_g(self, x, *args):
                     grad = lambda y: grad_g(y, *args) + (1 / 2) * (y - x)
                     return gradient_descent(grad, x, stepsize_inter, max_iter=max_iter_inter, tol=tol_inter)
