+++ deepinv/unfolded/deep_equilibrium.py
@@ -3,7 +3,7 @@ import torch.nn as nn
 from deepinv.optim.fixed_point import FixedPoint, AndersonAcceleration
 from deepinv.optim.optim_iterators import *
 from deepinv.unfolded.unfolded import BaseUnfold
-from deepinv.unfolded import str_to_class
+from deepinv.optim.utils import str_to_class
 
 class BaseDEQ(BaseUnfold):
     '''
