+++ deepinv/unfolded/unfolded.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 from deepinv.optim.fixed_point import FixedPoint, AndersonAcceleration
 from deepinv.optim.optim_iterators import *
-from deepinv.unfolded import str_to_class
+from deepinv.optim.utils import str_to_class
 from deepinv.optim.optimizers import BaseOptim
 
 class BaseUnfold(BaseOptim):
